server/requirements.txt
Flask
python-sqlite-orm
waitress
